Fix filtering for browser apps.
Despite what the javadoc on PackageManager.MATCH_ALL says, it actually
still performs all kinds of filtering and only affects whether to
return only the default browser for browser intent. So no need for
extra filtering on the result.

The previous filtering based upon ActivityInfo.enabled and
ApplicationInfo.enabled is incorrect because they only reflect the
state declared in manifest, not the actual state.

Fixes: 137731581
Test: presubmit
Change-Id: I61a1dcd572d2ada01ec96c9399be38f66ad88bbb